Refines CTown_Tasks to look more like an actual package.
Includes three basic functions for now to  see if the structure works well.

initializeModel() creates all the necessary parameters and stores them in a dictionary. This is so that a new wntr model and simulation does not have to be run every time one wants to create a new plot.

createBasicPlot() creates a basic plot similar to one you would see in EPANET.

createDemandPatternPlot() creates a demand pattern plot. Shows pipes and junctions, as well as resevoirs and tanks by default by I plan to have this be changeable by the user.